remove mango dep
