+++ lib/client/src/account_fetchers.rs
@@ -30,16 +30,6 @@ pub async fn account_fetcher_fetch_anchor_account<T: AccountDeserialize>(
         .with_context(|| format!("deserializing anchor account {}", address))
 }
 
-// Can't be in the trait, since then it would no longer be object-safe...
-pub async fn account_fetcher_fetch_mango_account(
-    fetcher: &dyn AccountFetcher,
-    address: &Pubkey,
-) -> anyhow::Result<MangoAccountValue> {
-    let account = fetcher.fetch_raw_account(address).await?;
-    let data: &[u8] = &account.data();
-    MangoAccountValue::from_bytes(&data[8..])
-        .with_context(|| format!("deserializing mango account {}", address))
-}
 
 pub struct RpcAccountFetcher {
     pub rpc: RpcClientAsync,
