+++ lib/client/src/lib.rs
@@ -5,6 +5,7 @@ pub use util::*;
 
 mod account_fetcher_trait;
 mod account_fetchers;
+mod mango_account_fetchers;
 pub mod account_update_stream;
 pub mod chain_data;
 mod chain_data_fetcher;
